move wake compute
